Move snapshot code into separate module
