+++ src/invoke.rs
@@ -15,6 +15,8 @@ use stellar_contract_env_host::{
 
 use crate::strval::{self, StrValError};
 
+use crate::snapshot;
+
 #[derive(Subcommand, Debug)]
 enum Commands {
     Call,
@@ -55,78 +57,10 @@ pub enum Error {
     Xdr(#[from] XdrError),
     #[error("host")]
     Host(#[from] HostError),
+    #[error("snapshot")]
+    Snapshot(#[from] snapshot::Error),
     #[error("serde")]
     Serde(#[from] serde_json::Error),
-}
-
-pub mod snapshot {
-    use std::fs::File;
-
-    use super::{Error, HostError};
-    use stellar_contract_env_host::{
-        im_rc::OrdMap,
-        storage::SnapshotSource,
-        xdr::{LedgerEntry, LedgerKey, VecM},
-    };
-
-    pub struct Snap {
-        pub ledger_entries: OrdMap<LedgerKey, LedgerEntry>,
-    }
-
-    impl SnapshotSource for Snap {
-        fn get(&self, key: &LedgerKey) -> Result<LedgerEntry, HostError> {
-            match self.ledger_entries.get(key) {
-                Some(v) => Ok(v.clone()),
-                None => Err(HostError::General("missing entry")),
-            }
-        }
-        fn has(&self, key: &LedgerKey) -> Result<bool, HostError> {
-            Ok(self.ledger_entries.contains_key(key))
-        }
-    }
-
-    // snapshot_file format is the default serde JSON representation of VecM<(LedgerKey, LedgerEntry)>
-    pub fn read(input_file: &std::path::PathBuf) -> Result<OrdMap<LedgerKey, LedgerEntry>, Error> {
-        let mut res = OrdMap::new();
-
-        let mut file = match File::open(input_file) {
-            Ok(f) => f,
-            Err(e) => {
-                //File doesn't exist, so treat this as an empty database and the file will be created later
-                if e.kind() == std::io::ErrorKind::NotFound {
-                    return Ok(res);
-                }
-                return Err(Error::Io(e));
-            }
-        };
-
-        let state: VecM<(LedgerKey, LedgerEntry)> = serde_json::from_reader(&mut file)?;
-        res = state.iter().cloned().collect();
-
-        Ok(res)
-    }
-
-    pub fn commit(
-        mut new_state: OrdMap<LedgerKey, LedgerEntry>,
-        storage_map: &OrdMap<LedgerKey, Option<LedgerEntry>>,
-        output_file: &std::path::PathBuf,
-    ) -> Result<(), Error> {
-        //Need to start off with the existing snapshot (new_state) since it's possible the storage_map did not touch every existing entry
-        let file = File::create(output_file)?;
-        for (lk, ole) in storage_map {
-            if let Some(le) = ole {
-                new_state.insert(lk.clone(), le.clone());
-            } else {
-                new_state.remove(lk);
-            }
-        }
-
-        let vec_new_state: VecM<(LedgerKey, LedgerEntry)> =
-            new_state.into_iter().collect::<Vec<_>>().try_into()?;
-        serde_json::to_writer(&file, &vec_new_state)?;
-
-        Ok(())
-    }
 }
 
 impl Cmd {
